feat(fipe-history): automatically store vehicle search for logged-in users

+++ src/main/java/fipe/preco/preco_fipe/controller/UserController.java
@@ -33,7 +33,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<UserGetResponse> findById(@PathVariable Long id) {
+    public ResponseEntity<UserGetResponse> findById(@PathVariable Integer id) {
         var user = service.findById(id);
 
         var userGetResponse = mapper.toUserGetResponse(user);
